src/requirements.txt
Django==4.2.7
django-environ==0.10.0
djangorestframework==3.14.0
pillow
